lazy load datasource so it doesn't need connection upon startup
