+++ sessions/next24/books-genai-vertex-langchain4j/src/main/java/services/config/AppConfig.java
@@ -25,24 +25,24 @@ import com.zaxxer.hikari.HikariDataSource;
 @Configuration
 public class AppConfig {
 
-    @Autowired
-    private Environment environment;
+//    @Autowired
+//    private Environment environment;
 
-    @Bean
-    public HikariDataSource getDataSource() {
-        HikariConfig config = new HikariConfig();
-        HikariDataSource ds;
-        config.setJdbcUrl(environment.getProperty("spring.datasource.url"));
-        config.setUsername(environment.getProperty("spring.datasource.username"));
-        config.setPassword(environment.getProperty("spring.datasource.password"));
-        config.addDataSourceProperty("cachePrepStmts", "true");
-        config.addDataSourceProperty("prepStmtCacheSize", "250");
-        config.addDataSourceProperty("prepStmtCacheSqlLimit", "2048");
-        config.addDataSourceProperty("databaseName", "quickstart_db");
-        config.addDataSourceProperty("port", "5000");
-        ds = new HikariDataSource(config);
-        return ds;
-    }
+//    @Bean
+//    public HikariDataSource getDataSource() {
+//        HikariConfig config = new HikariConfig();
+//        HikariDataSource ds;
+//        config.setJdbcUrl(environment.getProperty("spring.datasource.url"));
+//        config.setUsername(environment.getProperty("spring.datasource.username"));
+//        config.setPassword(environment.getProperty("spring.datasource.password"));
+//        config.addDataSourceProperty("cachePrepStmts", "true");
+//        config.addDataSourceProperty("prepStmtCacheSize", "250");
+//        config.addDataSourceProperty("prepStmtCacheSqlLimit", "2048");
+//        config.addDataSourceProperty("databaseName", "quickstart_db");
+//        config.addDataSourceProperty("port", "5000");
+//        ds = new HikariDataSource(config);
+//        return ds;
+//    }
 
 //    @Bean
 //    RestTemplate restTemplate() {
